Update README.md
Note on Development Timeline

I received the POE file on January 28th, which gave me limited time to complete the project. Due to the tight deadline, I focused on implementing the core features and was unable to include all planned functionalities. Additionally, I did not have enough time to publish the app on Google Play Console or the Play Store. Despite these constraints, I ensured that the essential requirements were met to create a functional and user-friendly note-taking application.